+++ backend/src/main/java/com/study/mock_sock/modules/stocks/services/dto/StockDto.java
@@ -7,7 +7,7 @@ public record StockDto(
         String ticker,
         String name,
         String currency,
-        int currentPrice,
+        double currentPrice,
         String exchange
 ) {
     public static StockDto from(Stock stock) {
